refactor(proto): split state_container.rs into a categorically-organized module
Pure move, zero behavior change. The ~2090-line state_container.rs is now a
state_container/ directory module, mirroring the packet_processor 'pub(crate)
mod includes' convention (children do use super::includes::*):
mod.rs (struct + helpers + create + getters + includes + re-exports),
virtual_connections, channels, inbound_transfer, outbound_transfer,
rekey_and_groups, monitoring, transfer_error.

All external crate::proto::state_container::{...} paths still resolve; only
visibility change is get_primary_stream private->pub(super) (now cross-file).
build + clippy clean; messaging/p2p/file-transfer/reconnection (15 tests) green.
